b1.4.7.1
Do mid updates maybe??

- Removed lots of warnings in the code
- Added VelocityA (doesnt work yet)
- Better memory (floats instead of doubles)
- Added MotionF
- Fixed StepA falsing in the void
- Fixed BadPacketsA
- Added BadPacketsC
- Removed a way to make an ACD disabler

+++ src/main/java/me/omgpandayt/acd/listeners/TeleportListener.java
@@ -16,7 +16,7 @@ public class TeleportListener implements Listener {
 		if(playerData == null)return;
 		
 		playerData.sinceTeleportTicks = 0;
-		playerData.lastGroundY = e.getTo().getY();
+		playerData.lastGroundY = (float)e.getTo().getY();
 		
 	}
